Switch to `forward_plus` renderer (read desc)
Godot will automatically fallback to compat renderer when needed:

'Since Godot 4.4, when using Forward+ or Mobile, if Vulkan is not
supported, the engine will fall back to Direct3D 12 and vice versa. If
the attempted fallback driver is not supported either, the engine will
then fall back to Compatibility when the RenderingDevice backend is not
supported. This allows the project to run anyway, but it may look
different than the intended appearance due to the more limited renderer.
This behavior can be disabled in the project settings by unchecking
Rendering > Rendering Device > Fallback to OpenGL 3.'

https://docs.godotengine.org/en/4.4/tutorials/rendering/renderers.html#switching-between-renderers